add, edit and delete student

+++ server-app/src/main/java/com/feut/entity/Role.java
@@ -13,7 +13,7 @@ public class Role {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.SEQUENCE)
-	private Long id;
+	private int id;
 	
 	@Column(name="name")
 	private String name;
@@ -25,13 +25,16 @@ public class Role {
 		
 	}
 
-	public Long getId() {
+
+	public int getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+
+	public void setId(int id) {
 		this.id = id;
 	}
+
 
 	public String getName() {
 		return name;
